mcp/pyproject.toml
[project]
name = "relay-mcp"
version = "0.1.0"
description = "MCP Server for Dedalus Labs multi-agent coordination"
requires-python = ">=3.10"
dependencies = [
    "dedalus-mcp",
    "httpx",
    "pydantic>=2.0.0",
    "python-dotenv",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
